controller created and routes modified
