+++ src/app.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import routes from "./routes/auth";
+import routes from "./routes/auth.route";
 
 const app = express();
 const port = 3000;
